Conflicts:
	admin/assets/js/generator.js.map
	admin/assets/js/generator.min.js

+++ public/includes/components/component-gallery.php
@@ -27,45 +27,34 @@ class AesopCoreGallery {
 		$atts 		= shortcode_atts($defaults, $atts);
 
 		// gallery ID
-		$gallery_id = $atts['id'];
+		$gallery_id = isset( $atts['id'] ) ? (int) $atts['id'] : false;
+
+		// alias to new atts type
 
 		// let this be used multiple times
 		static $instance = 0;
 		$instance++;
 		$unique 	= sprintf('%s-%s', $gallery_id, $instance);
 
-		// get the gallery
-		$gallery 	= wp_cache_get('aesop_gallery_retrieval_'.$unique);
-
-		// cache teh gallery retrieval
-		if ( false == $gallery ) {
-
-			$gallery 	= get_post_gallery( $gallery_id , false);
-			wp_cache_set('aesop_gallery_retrieval_'.$unique, $gallery);
-
-		}
-
 		// get gallery images and custom attrs
-		$image_ids 	= explode( ',', $gallery['ids'] );
-		$type 		= $gallery['a_type'];
+		$image_ids 	= get_post_meta($gallery_id,'_ase_gallery_images', true);
+		$image_ids	= array_map('intval', explode(',', $image_ids));
+
+		$type 		=  get_post_meta($gallery_id,'aesop_gallery_type', true);
 		$width 		= get_post_meta($gallery_id,'aesop_gallery_width', true);
 
 		//gallery caption
 		$gallery_caption = get_post_meta( $gallery_id, 'aesop_gallery_caption', true);
 
-		// set the type of gallery into post meta
-		// @todo - move this to  save_post action so it doesn't run every time the sc loads
-		update_post_meta( $gallery_id, 'aesop_gallery_type', $type );
-
 		ob_start();
 
 			do_action('aesop_gallery_before', $type, $gallery_id); //action
 
-			?><div id="aesop-gallery-<?php echo $unique;?>" class="aesop-component aesop-gallery-component aesop-<?php echo $type;?>-gallery-wrap"><?php
+			?><div id="aesop-gallery-<?php echo esc_attr( $unique );?>" <?php echo aesop_component_data_atts( 'gallery', $gallery_id, $atts );?> class="aesop-component aesop-gallery-component aesop-<?php echo esc_attr($type);?>-gallery-wrap <?php if ( empty( $gallery_id ) ) echo 'empty-gallery';?> "><?php
 
 				do_action('aesop_gallery_inside_top', $type, $gallery_id); //action
 
-				if ( !empty($gallery['ids']) ) {
+				if ( !empty($image_ids) ) {
 
 					switch($type):
 						case 'thumbnail':
@@ -89,20 +78,33 @@ class AesopCoreGallery {
 					endswitch;
 
 					if ( $gallery_caption ) {
-						printf('<p class="aesop-component-caption">%s</p>', $gallery_caption);
+						printf('<p class="aesop-component-caption">%s</p>', esc_html( $gallery_caption ) );
 					}
 
-					if ( is_user_logged_in() ) {
+					// provide the edit link to the backend edit if Aesop Editor is not active
+					if ( !class_exists('Lasso') && is_user_logged_in() && current_user_can('edit_posts') ) {
 						$url = admin_url( 'post.php?post='.$gallery_id.'&action=edit' );
 						$edit_gallery = __('edit gallery', 'aesop-core');
 						printf('<a class="aesop-gallery-edit aesop-content" href="%s" target="_blank" title="%s">(%s)</a>',$url, $edit_gallery, $edit_gallery );
 					}
 
-				} else {
+				}
 
-					?><div class="aesop-error aesop-content"><?php
-						_e('This gallery is empty! It\'s also possible that you simply have the wrong gallery ID.', 'aesop-core');
-					?></div><?php
+				if ( empty( $gallery_id ) && is_user_logged_in() && current_user_can('edit_posts') ) {
+
+					if ( class_exists('Lasso') ) {
+
+						?><div contenteditable="false" class="lasso--empty-component"><?php
+							_e('Setup a gallery by clicking the <span class="lasso-icon-gear"></span> icon above.', 'aesop-core');
+						?></div><?php
+
+					} else {
+
+						?><div class="aesop-error aesop-content"><?php
+							_e('This gallery is empty! It\'s also possible that you simply have the wrong gallery ID.', 'aesop-core');
+						?></div><?php
+
+					}
 				}
 
 				do_action('aesop_gallery_inside_bottom', $type, $gallery_id); //action
@@ -122,18 +124,18 @@ class AesopCoreGallery {
 	*/
 	function aesop_thumb_gallery($gallery_id, $image_ids, $width){
 
-		$thumbs 	= get_post_meta( $gallery_id, 'aesop_thumb_gallery_hide_thumbs', true) ? sprintf('data-nav="false"') : sprintf('data-nav="thumbs"');
+		$thumbs 	= get_post_meta( $gallery_id, 'aesop_thumb_gallery_hide_thumbs', true) ? sprintf('data-nav=false') : sprintf('data-nav=thumbs');
 		$autoplay 	= get_post_meta( $gallery_id, 'aesop_thumb_gallery_transition_speed', true) ? sprintf('data-autoplay="%s"', get_post_meta( $gallery_id, 'aesop_thumb_gallery_transition_speed', true)) : null;
-		$transition = get_post_meta( $gallery_id, 'aesop_thumb_gallery_transition', true) ? get_post_meta( $gallery_id, 'aesop_thumb_gallery_transition', true) : 'slide';
+		$transition = get_post_meta( $gallery_id, 'aesop_thumb_gallery_transition', true) ? get_post_meta( $gallery_id, 'aesop_thumb_gallery_transition', true) : 'crossfade';
 
 		// image size
 		$size    = apply_filters('aesop_thumb_gallery_size', 'full');
 
-		?><div id="aesop-thumb-gallery-<?php echo $gallery_id;?>" class="fotorama" 	data-transition="crossfade"
-																			data-width="<?php echo $width;?>"
-																			<?php echo $autoplay;?>
+		?><div id="aesop-thumb-gallery-<?php echo esc_attr($gallery_id);?>" class="fotorama" 	data-transition="<?php echo esc_attr($transition);?>"
+																			data-width="<?php echo esc_attr($width);?>"
+																			<?php echo esc_attr($autoplay);?>
 																			data-keyboard="true"
-																			<?php echo $thumbs;?>
+																			<?php echo esc_attr($thumbs);?>
 																			data-allow-full-screen="native"
 																			data-click="true"><?php
 
@@ -143,7 +145,7 @@ class AesopCoreGallery {
                 $alt     =  get_post_meta($image_id, '_wp_attachment_image_alt', true);
                 $caption 	= get_post($image_id)->post_excerpt;
 
-               ?><img src="<?php echo $full[0];?>" data-caption="<?php echo $caption;?>" alt="<?php echo esc_attr($alt);?>"><?php
+               ?><img src="<?php echo esc_url($full[0]);?>" data-caption="<?php echo esc_attr($caption);?>" alt="<?php echo esc_attr($alt);?>"><?php
 
 			endforeach;
 
@@ -157,8 +159,7 @@ class AesopCoreGallery {
 	*/
 	function aesop_grid_gallery($gallery_id, $image_ids, $width){
 
-		$getgridwidth 	= get_post_meta( $gallery_id, 'aesop_grid_gallery_width', true );
-		$gridwidth 		= $getgridwidth ? self::sanitize_int($getgridwidth) : 400;
+		$gridwidth 	= get_post_meta( $gallery_id, 'aesop_grid_gallery_width', true ) ? get_post_meta( $gallery_id, 'aesop_grid_gallery_width', true ) : 400;
 
 		// allow theme developers to determine the spacing between grid items
 		$space = apply_filters('aesop_grid_gallery_spacing', 5 );
@@ -170,19 +171,19 @@ class AesopCoreGallery {
 		<!-- Aesop Grid Gallery -->
 		<script>
 			jQuery(document).ready(function(){
-			    jQuery('#aesop-grid-gallery-<?php echo $gallery_id;?>').imagesLoaded(function() {
+			    jQuery('#aesop-grid-gallery-<?php echo esc_attr($gallery_id);?>').imagesLoaded(function() {
 			        var options = {
 			          	autoResize: true,
-			          	container: jQuery('#aesop-grid-gallery-<?php echo $gallery_id;?>'),
-			          	offset: <?php echo $space;?>,
-			          	flexibleWidth: <?php echo $gridwidth;?>
+			          	container: jQuery('#aesop-grid-gallery-<?php echo esc_attr($gallery_id);?>'),
+			          	offset: <?php echo (int) $space;?>,
+			          	flexibleWidth: <?php echo (int) $gridwidth;?>
 			        };
-			        var handler = jQuery('#aesop-grid-gallery-<?php echo $gallery_id;?> li');
+			        var handler = jQuery('#aesop-grid-gallery-<?php echo esc_attr($gallery_id);?> li');
 			        jQuery(handler).wookmark(options);
 			    });
 			});
 		</script>
-		<div id="aesop-grid-gallery-<?php echo $gallery_id;?>" class="aesop-grid-gallery aesop-grid-gallery" style="width:100%;max-width:<?php echo $width;?>;margin:0 auto;"><ul><?php
+		<div id="aesop-grid-gallery-<?php echo esc_attr($gallery_id);?>" class="aesop-grid-gallery aesop-grid-gallery" style="width:100%;max-width:<?php echo esc_attr($width);?>;margin:0 auto;"><ul><?php
 
 			foreach ($image_ids as $image_id):
 
@@ -191,7 +192,15 @@ class AesopCoreGallery {
                 $img_title 	  	= get_post($image_id)->post_title;
                 $caption 		= get_post($image_id)->post_excerpt;
 
-               	printf('<li class="aesop-grid-gallery-item"><a class="aesop-lightbox" href="%s" title="%s"><span class="aesop-grid-gallery-caption">%s</span><span class="clearfix">%s</span></a></li>',$getimagesrc[0], esc_attr($img_title), $caption, $getimage);
+			?>
+
+			<li class="aesop-grid-gallery-item"><a class="aesop-lightbox" href="<?php echo $getimagesrc[0];?>" title="<?php echo esc_attr($img_title);?>">
+				<?php if ( $caption ){ ?>
+					<span class="aesop-grid-gallery-caption"><?php echo $caption;?></span>
+				<?php } ?>
+			<span class="clearfix"><?php echo $getimage;?></span></a></li>
+
+			<?php
 
 			endforeach;
 
@@ -212,7 +221,7 @@ class AesopCoreGallery {
 		jQuery(document).ready(function(){
 
 			var stackedResizer = function(){
-				jQuery('.aesop-stacked-img').css({'height':(jQuery(window).height())+'px', 'width':(jQuery(this).parent())+'px'});
+				jQuery('.aesop-stacked-img').css({'height':(jQuery(window).height())+'px'});
 			}
 			stackedResizer();
 
@@ -235,9 +244,9 @@ class AesopCoreGallery {
             $caption 		= get_post($image_id)->post_excerpt;
 
            	?>
-           	<div class="aesop-stacked-img" style="background-image:url('<?php echo $full[0];?>');<?php echo $styles;?>">
+           	<div class="aesop-stacked-img" style="background-image:url('<?php echo esc_url($full[0]);?>');<?php echo $styles;?>">
            		<?php if ( $caption ){ ?>
-           			<div class="aesop-stacked-caption"><?php echo $caption;?></div>
+           			<div class="aesop-stacked-caption"><?php echo esc_html($caption);?></div>
            		<?php } ?>
            	</div>
            	<?php
@@ -256,26 +265,25 @@ class AesopCoreGallery {
 		// image size
 		$size    = apply_filters('aesop_sequence_gallery_size', 'large');
 
+		// lazy loader class
+		$lazy_holder =  AI_CORE_URL.'/public/assets/img/aesop-lazy-holder.png';
+
+
 		foreach ( $image_ids as $image_id ):
 
-            $img     =  wp_get_attachment_image($image_id, $size, false,'');
+            $img     =  wp_get_attachment_image_src($image_id, $size, false,'');
             $alt     =  get_post_meta($image_id, '_wp_attachment_image_alt', true);
             $caption 	= get_post($image_id)->post_excerpt;
+
+            $lazy   = class_exists('AesopLazyLoader') && !is_user_logged_in() ? sprintf( 'src="%s" data-src="%s" class="aesop-sequence-img aesop-lazy-img"',$lazy_holder, esc_url( $img[0] ) ) : sprintf( 'src="%s" class="aesop-sequence-img" ', esc_url( $img[0] ) );
 
            	?>
            	<figure class="aesop-sequence-img-wrap">
 
-           		<?php
+           		<img <?php echo $lazy;?> alt="<?php echo esc_attr($alt);?>">
 
-           		include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
-           		if (is_plugin_active('aesop-lazy-loader/aesop-lazy-loader.php')) {?>
-					<img class="aesop-sequence-img" data-original="<?php echo $img[0];?>" alt="<?php echo esc_attr($alt);?>">
-           		<?php } else {?>
-           			<img class="aesop-sequence-img" src="<?php echo $img[0];?>" alt="<?php echo esc_attr($alt);?>">
-           		<?php } ?>
-
-           		<?php if($caption){ ?>
-           			<figcaption class="aesop-content aesop-component-caption aesop-sequence-caption"><?php echo $caption;?></figcaption>
+           		<?php if ( $caption ) { ?>
+           			<figcaption class="aesop-content aesop-component-caption aesop-sequence-caption"><?php echo esc_html($caption);?></figcaption>
            		<?php } ?>
 
            	</figure>
@@ -298,7 +306,7 @@ class AesopCoreGallery {
 		// layout
 		$layout = get_post_meta( $gallery_id, 'aesop_photoset_gallery_layout', true) ? get_post_meta( $gallery_id, 'aesop_photoset_gallery_layout', true) : '';
 
-		$style 	= $width ? sprintf('style="max-width:%s;margin-left:auto;margin-right:auto;"', $width) : null;
+		$style 	= $width ? sprintf('style="max-width:%s;margin-left:auto;margin-right:auto;"', esc_attr( $width ) ) : null;
 
 		// lightbox
 		$lightbox = get_post_meta( $gallery_id, 'aesop_photoset_gallery_lightbox', true );
@@ -311,7 +319,7 @@ class AesopCoreGallery {
 		<script>
 			jQuery(window).load(function(){
 				jQuery('.aesop-gallery-photoset').photosetGrid({
-				  	gutter: "<?php echo $space.'px';?>",
+				  	gutter: "<?php echo absint($space).'px';?>",
 				  	<?php if ( $lightbox ) { ?>
 				  	highresLinks:true,
 				  	<?php } ?>
@@ -323,12 +331,12 @@ class AesopCoreGallery {
 				  		<?php } ?>
 
 					    jQuery('.aesop-gallery-photoset').attr('style', '');
-					    jQuery(".aesop-gallery-photoset img").each(function(){
+					    jQuery(".photoset-cell img").each(function(){
 
-							caption = jQuery(this).attr('alt');
+							caption = jQuery(this).attr('data-caption');
 
-							if ( caption ) {
-								title = jQuery(this).attr('data-title');
+							if ( caption) {
+								title = jQuery(this).attr('title');
 								jQuery(this).after('<span class="aesop-photoset-caption"><span class="aesop-photoset-caption-title">' + title + '</span><span class="aesop-photoset-caption-caption">' + caption +'</span></span>');
 								jQuery('.aesop-photoset-caption').hide().fadeIn();
 
@@ -344,17 +352,18 @@ class AesopCoreGallery {
 			echo '<div class="aesop-gallery-photoset-width" '.$style.' >';
 		}
 
-			?><div class="aesop-gallery-photoset" data-layout="<?php echo $layout;?>" ><?php
+			?><div class="aesop-gallery-photoset" data-layout="<?php echo absint($layout);?>" ><?php
 
 				foreach ( $image_ids as $image_id ):
 
 		            $full    	=  wp_get_attachment_image_src( $image_id, $size, false);
 		            $alt     	=  get_post_meta( $image_id, '_wp_attachment_image_alt', true );
+		            $caption    = 	get_post($image_id)->post_excerpt;
                 	$title 	  	= 	get_post($image_id)->post_title;
 
-		            $lb_link    =  $lightbox ? sprintf('data-highres="%s"', $full[0]) : null;
+		            $lb_link    =  $lightbox ? sprintf('data-highres="%s"', esc_url( $full[0] ) ) : null;
 
-		           	?><img src="<?php echo $full[0];?>" <?php echo $lb_link;?> data-title="<?php echo $title;?>" title="<?php echo $title;?>" alt="<?php echo $alt;?>"><?php
+		           	?><img src="<?php echo esc_url($full[0]);?>" <?php echo $lb_link;?> data-caption="<?php echo esc_attr($caption);?>" title="<?php echo esc_attr($title);?>" alt="<?php echo esc_attr($alt);?>"><?php
 
 				endforeach;
 
@@ -366,14 +375,5 @@ class AesopCoreGallery {
 
 	}
 
-    /**
-	 	* Ensure users only enter whole number
-	 	*
-	 	* @since    1.0.0
-	*/
-	function sanitize_int( $input = ''  ) {
-		return wp_filter_nohtml_kses( round( $input ) );
-
-	}
 }
 new AesopCoreGallery;
